requirements.txt
Pillow>=9.4.0
python-Levenshtein>=0.20.9
vaik-text-recognition-tflite-inference @ git+https://github.com/vaik-info/vaik-text-recognition-tflite-inference.git@7073a39a91e85cb8660e39a1c4a1e15b373d000b